Fix some issues and improve code
