+++ dear-diary/src/components/Routes/routes.tsx
@@ -6,15 +6,20 @@ interface RouteConfig {
   path: string;
   element: React.ReactNode;
 }
+
+export const paths = {
+  SignIn: "/",
+  HomePage: "/home",
+};
 const routes: RouteConfig[] = [
   {
     id: 1,
-    path: "/",
+    path: paths.SignIn,
     element: <SignIn />,
   },
   {
     id: 2,
-    path: "/home",
+    path: paths.HomePage,
     element: <ProtectedHome />,
   },
 ];
